requirements.txt
flask==3.0.0
torch>=2.1.0
transformers>=4.44.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
numpy>=1.26.0
